backend/requirements.txt
flask
gunicorn
supabase>=2.0.0
python-dotenv
python-dotenv